+++ cmd/repo-updater/repos/types.go
@@ -948,7 +948,10 @@ func NewRateLimiterRegistry(ctx context.Context, store Store) (*RateLimiterRegis
 	for _, svc := range svcs {
 		err = r.updateRateLimiter(svc)
 		if err != nil {
-			// Errors here are not fatal
+			if _, ok := err.(errRateLimitUnsupported); ok {
+				continue
+			}
+			// Errors here are not fatal, so we can log them
 			log15.Warn("Updating rate limiter", "kind", svc.Kind, "err", err)
 		}
 	}
@@ -1041,11 +1044,19 @@ func (r *RateLimiterRegistry) updateRateLimiter(svc *ExternalService) error {
 			}
 		}
 	default:
-		return fmt.Errorf("internal rate limiting not support for %s", svc.Kind)
+		return errRateLimitUnsupported{codehostKind: svc.Kind}
 	}
 
 	l := r.GetRateLimiter(svc.ID)
 	l.SetLimit(limit)
 
 	return nil
+}
+
+type errRateLimitUnsupported struct {
+	codehostKind string
+}
+
+func (e errRateLimitUnsupported) Error() string {
+	return fmt.Sprintf("internal rate limiting not supported for %s", e.codehostKind)
 }